// frontend/src/app/core/generated/circabc/model/restoreNodeMetadata.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RestoreNodeMetadata { 
    /**
     * the node Id of the archive to be restored
     */
    archiveNodeId?: string;
    /**
     * he node Id of the target space into which the node will be restored it can be left empty 
     */
    targetFolderId?: string;
}